+++ packages/sanity/src/core/releases/contexts/upsell/types.ts
@@ -9,7 +9,7 @@ export interface ReleasesUpsellContextValue {
   mode: 'upsell' | 'default' | 'disabled'
   upsellDialogOpen: boolean
   guardWithReleaseLimitUpsell: (callback: () => void, throwError?: boolean) => false | void
-  onReleaseLimitReached: (limit: number) => void
+  onReleaseLimitReached: (limit: number, suppressDialogOpening: boolean) => void
   telemetryLogs: {
     dialogSecondaryClicked: () => void
     dialogPrimaryClicked: () => void
